dont round the numbers
the reason i decided to round them in the first place was because i
calculated the GDP per capita using the two “preloaded” datasets i’ve
had and that resulted in many float values which were too long. but now
that i’ve gotten rid of that calculated dataset, i can get rid of this
unnecessary restriction.